requirements.txt
jupyter>=1.0.0
transformers
accelerate
torchinfo
einops
scipy
webdataset
sentencepiece
wandb[media]
torchmetrics[image]
simple_parsing
# decord   # ← Python 3.14 + M5 だとビルド済みパッケージが無いので一旦オフ
opencv-python
psutil
git+https://github.com/KeKsBoTer/torch-dwt
git+https://github.com/huggingface/diffusers.git
pyyaml
# av==12.3.0   # とりあえず動画用は後回し。画像だけなら不要

